+++ app/test/features/attachments/widget/file_attachment_preview_test.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:acter/features/attachments/widgets/file_attachment_preview.dart';
+import 'package:acter/features/attachments/widgets/attachment_preview/file_attachment_preview.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
